bin/ingest/requirements.txt
requests==2.21.0
pandas==1.1.5
python-dotenv==1.0.0
